+++ src/DependencyInjection/Configuration.php
@@ -43,7 +43,7 @@ final class Configuration implements ConfigurationInterface
                             ->addDefaultsIfNotSet()
                             ->children()
                                 ->scalarNode('version')
-                                    ->defaultValue('0.1.2')
+                                    ->defaultValue('0.2.0')
                                     ->cannotBeEmpty()
                                     ->validate()
                                         ->ifTrue(static fn (mixed $value): bool => !is_string($value) || preg_match('/^\d+\.\d+\.\d+(?:[-+][0-9A-Za-z.-]+)?$/', $value) !== 1)